refactor(command binding): Refactored command binding examples with simulated data sources for m...
# Project State

## Current Focus
Refactored command binding examples with simulated data sources for more reliable testing

## Context
The command binding examples were previously using real shell commands which could fail or produce inconsistent output. This change replaces them with simulated data sources that provide predictable, controlled output for testing and demonstration purposes.

## Completed
- [x] Replaced all external command executions with simulated data generation
- [x] Simplified widget initialization with default values
- [x] Improved testability by removing external dependencies
- [x] Added more realistic simulated data patterns for each widget type

## In Progress
- [ ] None

## Blockers
- None

## Next Steps
1. Update documentation to reflect the new simulated data approach
2. Add configuration options for test scenarios with different data patterns
---
category: refactor
scope: command binding
---

+++ examples/_cookbook/debug_overlay.rs
@@ -84,7 +84,7 @@ impl DebugOverlayPanel {
             id: WidgetId::new(10),
             label: focused_widget.unwrap_or("List_1").to_string(),
             children: (0..20).map(|i| WidgetNode {
-                id: WidgetId::new(11 + i as u64),
+                id: WidgetId::new(11 + i as usize),
                 label: format!("Item_{}", i + 1),
                 children: vec![],
             }).collect(),
